# Use multi_get for store that has extended API support.