--- a/gloo/common/utils.cc
+++ b/gloo/common/utils.cc
@@ -36,4 +36,10 @@ bool useRankAsSeqNumber() {
       (std::string(res) == "True" || std::string(res) == "1");
 }
 
+bool isStoreExtendedApiEnabled() {
+  const auto& res = std::getenv("GLOO_ENABLE_STORE_V2_API");
+  return res != nullptr &&
+      (std::string(res) == "True" || std::string(res) == "1");
+}
+
 } // namespace gloo