--- a/gloo/common/utils.h
+++ b/gloo/common/utils.h
@@ -16,4 +16,6 @@ std::string getHostname();
 
 bool useRankAsSeqNumber();
 
+bool isStoreExtendedApiEnabled();
+
 } // namespace gloo