--- a/gloo/transport/tcp/listener.cc
+++ b/gloo/transport/tcp/listener.cc
@@ -15,6 +15,7 @@
 #include <gloo/common/logging.h>
 #include <gloo/common/utils.h>
 #include <gloo/transport/tcp/helpers.h>
+#
 
 namespace gloo {
 namespace transport {