Summary:
The TCP store has API v2 support we can reduce the network overhead of Gloo rendezvous significantly by fetching a batch of key instead of doing them one by one.
Initial testing shows ~15X improvement for 4k jobs.

Gloo process group init:

Baseline ( fbcode trunk):
2k job (https://fburl.com/mlhub/x1prxu89) : ~82sec (~1.4 min)
4k job (https://fburl.com/mlhub/v1djk4n5) : ~393 sec (~6.6min)
8k job (https://fburl.com/mlhub/cagqrs7m): (~55mins)

With optimizations (D48130088 + D52083376):

2k job (https://fburl.com/mlhub/x0cskdag)  : ~18 sec ( ~5x faster)
4k job (https://fburl.com/mlhub/xzmvkm4j) : ~ 25 sec (~15x faster)
8k job (https://fburl.com/mlhub/gdyeizv9)   : ~ 85 sec (~35x faster)

Reviewed By: xunnanxu

Differential Revision:
D52083376

Privacy Context Container: L1156430